requirements.txt
torch==2.0.0
torchvision==0.15.1
datasets==2.10.1
evaluate==0.4.0
transformers==4.27.3
sentencepiece==0.1.97
timm==0.6.13
seqeval==1.2.2
sentence-transformers==2.2.2
tensorflow-cpu==2.12.0
